Add `SNPs.merge` method; obsolete `SNPsCollection` and various output files
